refactor: rename dashboard variable related components
